requirements.txt
Flask==3.0.0
python-ldap==3.4.4
# libtorrent==2.0.11 - Installation depuis GitHub (compile automatiquement)
# pip install git+https://github.com/arvidn/libtorrent.git@v2.0.11#subdirectory=bindings/python
passlib==1.7.4
requests==2.31.0
Flask-WTF==1.2.1
Flask-Talisman==1.1.0
Flask-Compress==1.14
Flask-Limiter==3.5.0
redis==5.0.1
